change(common/models): baseline es module conversion of lm-layer
