--- common/predictive-text/browser.tsconfig.json
@@ -1,31 +1,0 @@
-{
-  "extends": "../../tsconfig-base.json",
-
-  "compilerOptions": {
-    "allowJs": true,
-    "module": "none",
-    "outFile": "./build/index.js",
-    "inlineSources": true,
-    "sourceMap": true,
-    "target": "es5",
-    "lib": ["es6", "dom"]
-  },
-  "include" : ["./*.ts"],
-  "references": [
-    { "path": "../models/types" },
-    { "path": "../models/wordbreakers" },
-    { "path": "../models/templates" },
-    { "path": "../web/lm-message-types" }
-  ],
-  "exclude" : [
-    "node-defaultWorker.ts",
-    "virtualizedWorker.ts",
-    "index.d.ts",
-    "worker-interface.d.ts",
-    "node_modules",
-    "testing"
-  ],
-  "files": [
-    "../web/lm-worker/build/lib/worker-main.wrapped-for-bundle.ts"
-  ]
-}
