--- common/predictive-text/embedded_worker.d.ts
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2018 National Research Council Canada (author: Eddie A. Santos)
- * Copyright (c) 2018 SIL International
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-// Include the code intended to run WITHIN the Web Worker.
-// The worker code MUST be compiled before this file is compiled.
-// If you see a 'File: embedded_worker.js not found.' error message, please
-// compile the worker first (stage one).
-/*/// <reference path="build/intermediate/embedded_worker.js" />*/
-
-/**
- * The body of this function contains code that should run
- * inside a Web Worker context.
- *
- * Convert this code into a blob URI with LMLayer.asBlobURI(),
- * and pass the URI into the new Worker() constructor.
- */
- declare function LMLayerWorkerCode(): void;
