+++ common/predictive-text/src/index.ts
@@ -1,0 +1,3 @@
+export { default as LMLayer } from './lmlayer.js';
+export { default as NodeWorker } from './node/worker.js';
+export { default as WebWorker } from './web/worker.js';
