+++ common/predictive-text/src/node/tsconfig.json
@@ -1,0 +1,26 @@
+{
+  "extends": "../../../../tsconfig-base.json",
+  "compilerOptions": {
+    "allowJs": false,
+    "declaration": true,
+    "module": "es6",
+    "moduleResolution": "Node",
+    "inlineSources": true,
+    "sourceMap": true,
+    "sourceRoot": "keyman",
+    "target": "es5",
+    "lib": ["es6"],
+    "types": ["node"],
+    "baseUrl": "../",
+    "outDir": "../../build/obj",
+    "tsBuildInfoFile": "../../build/obj/node/tsconfig.tsbuildinfo",
+    "rootDir": "../"
+  },
+  "include": [
+    "*.ts"
+  ],
+  "references": [
+    { "path": "../tsconfig.json" },
+    { "path": "../../../web/lm-worker" }
+  ]
+}
