+++ common/predictive-text/src/tsconfig.json
@@ -1,0 +1,33 @@
+{
+  "extends": "../../../tsconfig-base.json",
+
+  "compilerOptions": {
+    "allowJs": false,
+    "declaration": true,
+    "module": "es6",
+    "moduleResolution": "Node",
+    "inlineSources": true,
+    "sourceMap": true,
+    "sourceRoot": "keyman",
+    "target": "es5",
+    "lib": ["es6"],
+    "types": ["node"],
+    "baseUrl": "./",
+    "outDir": "../build/obj",
+    "tsBuildInfoFile": "../build/obj/tsconfig.tsbuildinfo",
+    "rootDir": "./"
+  },
+  "references": [
+    { "path": "../../web/utils" },
+    { "path": "../../models/types"},
+    { "path": "../../models/wordbreakers"},
+    { "path": "../../web/lm-message-types"},
+    { "path": "../../models/templates"}
+  ],
+  "include" : [ "./*.ts" ],
+  "exclude" : [
+    "node",
+    "web",
+    "index.ts" // It's handled by the final layer.
+  ]
+}
