+++ common/predictive-text/src/web/tsconfig.json
@@ -1,0 +1,25 @@
+{
+  "extends": "../../../../tsconfig-base.json",
+  "compilerOptions": {
+    "allowJs": false,
+    "declaration": true,
+    "module": "es6",
+    "moduleResolution": "Node",
+    "inlineSources": true,
+    "sourceMap": true,
+    "sourceRoot": "keyman",
+    "target": "es5",
+    "lib": ["es6", "DOM"],
+    "types": ["node"],
+    "baseUrl": "../",
+    "outDir": "../../build/obj",
+    "tsBuildInfoFile": "../../build/obj/web/tsconfig.tsbuildinfo",
+    "rootDir": "../"
+  },
+  "include": [
+    "*.ts"
+  ],
+  "references": [
+    { "path": "../tsconfig.json" }
+  ]
+}
