+++ common/predictive-text/src/worker-interface.d.ts
