--- common/predictive-text/tsconfig.json
@@ -1,32 +1,0 @@
-{
-  "extends": "../../tsconfig-base.json",
-
-  "compilerOptions": {
-    "allowJs": true,
-    "outFile": "build/headless.js",
-    "module": "none",
-    "inlineSources": true,
-    "sourceMap": true,
-    "target": "es5",
-    "lib": ["es6"]
-  },
-  "references": [
-    { "path": "../web/utils" },
-    { "path": "../models/types"},
-    { "path": "../models/wordbreakers"},
-    { "path": "../web/lm-message-types"},
-    { "path": "../models/templates"},
-  ],
-  "include" : [ "./*.ts" ],
-  "exclude" : [
-    "web-defaultWorker.ts",
-    "node_modules",
-    "testing"
-  ],
-  "files": [
-    // By renaming the output JS as TS, and including the @ts-nocheck directive, we can
-    // bypass wrapped-worker typing issues while still bundling the file as part of the
-    // tsc build!
-    "../web/lm-worker/build/lib/worker-main.wrapped-for-bundle.ts"
-  ]
-}
