+++ common/web/lm-worker/build.sh
@@ -22,7 +22,7 @@ cd "$(dirname "$THIS_SCRIPT")"
 
 WORKER_OUTPUT=build/obj
 WORKER_OUTPUT_FILENAME=build/lib/worker-main.js
-WORKER_WRAPPED_BUNDLE_TARGET_FILENAME=build/lib/worker-main.wrapped-for-bundle.ts
+WORKER_WRAPPED_BUNDLE_TARGET_FILENAME=build/lib/worker-main.wrapped-for-bundle.js
 
 # Wraps JavaScript code in a way that can be embedded in a worker.
 # To get the inner source code, include the file generated by this function,
@@ -37,7 +37,7 @@ wrap-worker-code ( ) {
   # Use of this annotation allows us to actually build against the file, rather
   # than hackily insert it later when/where needed.
   echo "// @ts-nocheck"
-  printf "function %s () {\n" "${name}"
+  printf "export default function %s () {\n" "${name}"
 
   # Since the worker is compiled with "allowJS=false" so that we can make
   # declaration files, we have to insert polyfills here.
